Split parser tests
